ReflectSharp
It's been a long time The reason for the lack of activity is not only studying C language, but also studying various things, and these days, I am being treated for injuries, and I am working separately, so I did not have enough time. Let's hang in there for a month.
And this is the ReflectSharp task in the book. I don't know how it works because I haven't found an example. I tried to make it as much as I could.